add go library for common functions for integration tests implemented as go_bazel_test
