+++ tests/BUILD.bazel
@@ -25,6 +25,10 @@ load(
     "//:constants.bzl",
     "test_ghc_version",
 )
+load(
+    "@io_bazel_rules_go//go:def.bzl",
+    "go_library"
+)
 
 exports_files(
     [
@@ -415,6 +419,14 @@ haskell_repl(
         "//tests:run-tests",
         "//tests/binary-with-lib",
     ],
+)
+
+go_library(
+    name = "integration_testing",
+    srcs = ["integration_testing.go"],
+    deps = ["@io_bazel_rules_go//go/tools/bazel_testing"],
+    importpath = "github.com/tweag/rules_haskell/tests/integration_testing",
+    visibility = ["//visibility:public"],
 )
 
 filegroup(
