+++ tests/repl-targets/hs_bin_repl_test.go
@@ -1,19 +1,12 @@
 package hs_bin_repl_test
 
 import (
-        "bytes"
-        "fmt"
-        "os"
-        "os/exec"
-        "runtime"
         "testing"
-        "github.com/bazelbuild/rules_go/go/tools/bazel_testing"
+        it "github.com/tweag/rules_haskell/tests/integration_testing"
 )
 
-
-
-func Workspace() string {
-        return `
+func TestMain(m *testing.M) {
+        it.TestMain(m, `
 -- WORKSPACE --
 local_repository(
         name = "rules_haskell",
@@ -98,67 +91,13 @@ module QuuxLib (message) where
 
 message :: String
 message = "Hello GHCi!"
-`
-}
-
-func TestMain(m *testing.M) {
-        bazel_testing.TestMain(m, bazel_testing.Args{
-                Main: Workspace() + GenerateBazelrc(),
-        })
-}
-
-func AssertOutput(t *testing.T, output []byte, expected string) {
-        if string(output) != expected {
-                t.Fatalf("output of bazel process is invalid.\nExpected: %v\n, Actual: %v\n", expected, string(output))
-        }
-}
-
-func UseNixpkgs() bool {
-        for _, arg := range os.Args {
-                if arg == "nixpkgs=true" {
-                        return true
-                }
-        }
-        return false
-}
-
-func GenerateBazelrc() string {
-        bazelrc := "-- .bazelrc --\n"
-        if UseNixpkgs() {
-                bazelrc += "build --host_platform=@io_tweag_rules_nixpkgs//nixpkgs/platforms:host\n"
-        } else if runtime.GOOS == "windows" {
-                bazelrc += "build --crosstool_top=@rules_haskell_ghc_windows_amd64//:cc_toolchain\n"
-        }
-        return bazelrc
-}
-
-func BazelOutput(args ...string) ([]byte, error) {
-        cmd := bazel_testing.BazelCmd(args...)
-
-        // It's important value of $HOME to be invariant between different integration test runs
-        // and to be writable directory for bazel test. Probably TEST_TMPDIR is a valid choice
-        // but documentation is not clear about it's default value
-        // cmd.Env = append(cmd.Env, fmt.Sprintf("HOME=%s", os.Getenv("TEST_TMPDIR")))
-        cmd.Env = append(cmd.Env, fmt.Sprintf("HOME=%s", os.TempDir()))
-        if runtime.GOOS == "darwin" {
-                cmd.Env = append(cmd.Env, "BAZEL_USE_CPP_ONLY_TOOLCHAIN=1")
-        }
-        stdout := &bytes.Buffer{}
-        stderr := &bytes.Buffer{}
-        cmd.Stdout = stdout
-        cmd.Stderr = stderr
-        err := cmd.Run()
-        if eErr, ok := err.(*exec.ExitError); ok {
-                eErr.Stderr = stderr.Bytes()
-                err = &bazel_testing.StderrExitError{Err: eErr}
-        }
-        return stdout.Bytes(), err
+`)
 }
 
 func TestHsBinRepl(t *testing.T) {
-        out, err := BazelOutput("run", "//:hs-bin@repl", "--", "-ignore-dot-ghci", "-e", ":main")
+        out, err := it.BazelOutput("run", "//:hs-bin@repl", "--", "-ignore-dot-ghci", "-e", ":main")
         if err != nil {
                 t.Fatal(err)
         }
-        AssertOutput(t, out, "Hello GHCi!\n")
+        it.AssertOutput(t, out, "Hello GHCi!\n")
 }
